✨ 新赛季排行榜 API 模型

+++ nonebot_plugin_tetris_stats/games/tetrio/api/schemas/summaries/zenith.py
@@ -76,7 +76,7 @@ class Stats(BaseModel):
     kills: int
     finesse: Finesse
     zenith: _Zenith
-    finaltime: int
+    finaltime: float
 
 
 class Results(BaseModel):
